+++ include/wma/backends/glfw/GlfwWindowManager.hpp
@@ -33,6 +33,7 @@ namespace wma {
 
         void createWindow(const char* windowName) override;
         void pollEvents() override;
+        void waitEvents(int timeoutMs) override;
         void swapBuffers() override;
         void* getWindowInstance() override;
         void* getGLProcAddress(const char* name) const override;